% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_RestPoLL_fonts.r
\name{import_fonts}
\alias{import_fonts}
\title{Import Times New Roman and Arial fonts for use in charts}
\usage{
import_fonts()
}
\description{
These fonts are used when the AGaramondPro and FrutigerLTStd are not possible
to use.
}
\note{
You will need to load the fonts aswell with extrafont::loadfonts(),
this is done automatically when using the save_poms_plot().This will take care of ensuring PDF/PostScript usage. The location of
the font directory is displayed after the base import is complete. It is
highly recommended that you install them on your system the same way you
would any other font you wish to use in other programs.
}
