% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{LU_cols}
\alias{LU_cols}
\title{Function to extract Lund university colours as hex codes}
\usage{
LU_cols(...)
}
\arguments{
\item{...}{Character names of Lu_colours}
}
\description{
Function to extract Lund university colours as hex codes
}
