% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestPoLL_pal.r
\name{RestPoLL_pal}
\alias{RestPoLL_pal}
\title{Shows the RestPoLL Palette Colours}
\usage{
RestPoLL_pal(palette = "both")
}
\arguments{
\item{palette}{One of the PoMS colour palettes: "main", "secondary", "both", "dark", "grey", or "all".}
}
\value{
A colour map with respective Hex names
}
\description{
Show the colours with the Hex colour names of the given palette.
}
\examples{
RestPoLL_pal(palette = "main")

}
