% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{restpoll_pals}
\alias{restpoll_pals}
\title{Return function to interpolate a RestPoLL colour palette}
\usage{
restpoll_pals(palette = "both", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in LU_palettes ("main", "secondary",
"both").}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments to pass to colorRampPalette().}
}
\description{
Return function to interpolate a RestPoLL colour palette
}
