% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_RestPoLL_plot.R
\name{save_restpoll_plot}
\alias{save_restpoll_plot}
\title{Arrange, style and save the ggplot chart}
\usage{
save_restpoll_plot(
  save_filepath = "FigX.png",
  plot = last_plot(),
  source_name =
    "This project receives funding from the European Union's\\nHorizon Europe Framework Programme under project No. 101082101.",
  width = 9,
  height = 6,
  logo_image_path = file.path(system.file("logos", package = "RestPoLLggtheme"),
    "EU-flag.png"),
  RestPoLL_logo = "placeholderL.png"
)
}
\arguments{
\item{save_filepath}{Exact file-path that you want the plot to be saved to,
or just the name of the file. Default file-path is the working directory}

\item{plot}{The name of the plot (saved in an object) you have created that
you want to format and save. By default it saves the last plot made.}

\item{source_name}{The text you want to come after the text 'Source:' in the
bottom left hand side of your side. Defaults to the funding text of EU
Horizon project}

\item{width}{Width in inches that you want to save your chart to - defaults
to 9.}

\item{height}{Height in inches that you want to save your chart to - defaults
to 6.}

\item{logo_image_path}{File path for the logo image you want to use in the
bottom left of your chart, which needs to be a PNG file - defaults to a
EU-flag image.}

\item{RestPoLL_logo}{File path to the top right logo - defaults to a
placeholder image of background colour. There are different logos in colour
and black and white. However, you can not add a path to your own Logo. Use
\code{logo_image_path} for that. See \code{Section RestPoLL Logos} for
possible options.}
}
\value{
A png file with a styled and updated ggplot chart. Also returns the
  plot object.
}
\description{
Running this function with the plot name as variable will save your plot with some
extra styling. It will add Arial as default text font. It will also left align your
title, subtitle and source, add and optional RestPoLL logo in the top right, and save
it to your specified location.
}
\details{
The function is a modified version of a finalise_plot function from
 the bbplot package made by the BBC data team.
 \url{https://github.com/bbc/bbplot}
}
\section{RestPoLL logos}{

 Here is the list of possible choices for the RestPoLL logo:
 \itemize{
   \item Restpoll.png - The RestPoLL logo in colour, text under.
   \item Restpoll-bw.png - The RestPoLL logo in black and white, text under.
   \item Restpoll-long.png - The RestPoLL logo in colour, text to the right.
   \item Restpoll-long-bw.png - The RestPoLL logo in black and white, text to the right.
   \item Restpoll-notext.png - The RestPoLL logo in colour, without text.
   \item Restpoll-notext-bw.png - The RestPoLL logo in black and white, without text.
    }
}

\examples{
\dontrun{
save_restpoll_plot(save_filepath = "filename_that_my_plot_should_be_saved_to.png",
             plot = last_plot(),
             source_name = "Source: The source for my data",
             width = 9,
             height = 6,
             logo_image_path = "my_logo.png",
             RestPoLL_logo = "Restpoll.png")

library(ggplot2)
myplot <- ggplot(iris, aes(Petal.Length, Sepal.Length)) +
   geom_point() +
   facet_wrap(~Species) +
   theme_RestPoLL()

save_restpoll_plot(save_filepath = "Iris_facets.png",
             plot = myplot,
             source_name = "Source: Iris data",
             width = 9,
             height = 6)

}

}
\keyword{save_restpoll_plot}
