% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{scale_colour_RestPoLL}
\alias{scale_colour_RestPoLL}
\title{Color Scale Constructor for RestPoLL Colors}
\usage{
scale_colour_RestPoLL(
  palette = "secondary",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in RestPoLL palettes ("main", "secondary",
"both"). Show colours of the palettes with `RestPoLL_pal()`.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color Scale Constructor for RestPoLL Colors
}
\details{
The palette is made up of the different colors from the graphical profile of
  the RestPoLL project. The choices are:
\itemize{
   \item main - the main blue and bronze colours (can be used for first and secondary headings).
   \item secondary - the colours that are not main or grey.
   \item both - both the main and secondary colours.
 }
}
