% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{scale_fill_RestPoLL}
\alias{scale_fill_RestPoLL}
\title{Fill Scale constructor for RestPoLL Colours}
\usage{
scale_fill_RestPoLL(
  palette = "secondary",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in RestPoLL palettes ("main", "secondary",
"both"). Show colours of the palettes with `RestPoLL_pal()`.}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
Fill Scale constructor for RestPoLL Colours
}
