% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_RestPoLL.R
\name{theme_RestPoLL}
\alias{theme_RestPoLL}
\title{Add a RestPoLL Theme to ggplot2 Figures}
\usage{
theme_RestPoLL(base_size = 22, base_family = "Arial",
   title_size = 28, title_family = "Arial", subtitle_size = 24,
   subtitle_family = "Arial")
}
\arguments{
\item{base_size}{Font size}

\item{base_family}{Font family for the basic text.}

\item{title_size}{Font size of plot title.}

\item{title_family}{Font family for the titles. Including plot title, as
well as the facet strip titles. However, not for subtitle.}

\item{subtitle_size}{Font size of plot subtitle.}

\item{subtitle_family}{Font family for the subtitle.}
}
\value{
Returns the plot object made with your \link{ggplot2} functions but styled for Pollinator monitoring at LU
}
\description{
This function allows you to add the RestPoLL theme to your ggplot graphics.
}
\details{
\code{theme_RestPoLL} sets the theme for RestPoLL figures,
  using the graphical profile colours in the figure plot area, gridline,
  axis, facets etc. You can change the size and font family of the basic text, title and subtitle
  directly in the theme function. The article theme is pre-set to suit articles with a bold title in bronze colour.
}
\examples{
   library(ggplot2)
   library(RetPoLLggtheme)

   p1 <- ggplot(diamonds, aes(carat, price, color = color)) +
     geom_point() +
     facet_wrap(~cut) +
     labs(title = "Test of the Lund theme", subtitle = "I hope it works well")

   p1 + theme_RestPoLL()

}
\seealso{
\code{\link{ggplot2}}
}
\author{
Georg Andersson, Biodiversity, Lund University
}
\keyword{theme_PoMS}
