% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{Load Fonts For the R Session}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{Not used}

\item{pkgname}{Not used}
}
\value{
Load fonts for widows device and pdf device
}
\description{
Attach the installed fonts for this R session for the win and pdf devices.
}
