% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_RestPoLL_fonts.r
\name{import_fonts}
\alias{import_fonts}
\title{Import Times New Roman, Arial, and Lora fonts for use in charts}
\usage{
import_fonts()
}
\description{
These default fonts are Arial, however the Lora font is the recomended font for RestPoLL.
}
\note{
You should not need to load the fonts (e.g. with \code{extrafont::loadfonts()}).
This should be done with the \code{showtext} package. The location of
the font directory is displayed after the base import is complete. It is
highly recommended that you install them on your system the same way you
would any other font you wish to use in other programs.
}
