% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{scale_colour_RestPoLL}
\alias{scale_colour_RestPoLL}
\title{Color scale constructor for PoMS  colors}
\usage{
scale_colour_RestPoLL(
  palette = "secondary",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in LU_palettes ("main", "secondary",
"both", "grey", "dark", "all").}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\description{
Color scale constructor for PoMS  colors
}
\details{
The palette is made up of the different colors from the graphical profile of Lund university. The choices are:
\itemize{
   \item main - the main blue and bronze colours (can be used for first and secondary headings).
   \item secondary - the colours that are not main or grey.
   \item both - both the main and secondary colours.
   \item grey - The grey colours used by Lund university (mainly for headings and, ingress and main texts)
   \item dark - The blue, bronze and darker shade of grey.
   \item all - Use all the colours (default).
 }
}
