% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colours_RestPoLL.R
\name{scale_fill_RestPoLL}
\alias{scale_fill_RestPoLL}
\title{Fill scale constructor for PoMS colours}
\usage{
scale_fill_RestPoLL(
  palette = "secondary",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in LU_palettes ("main", "secondary",
"both", "grey", "dark", "all").}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not.}

\item{reverse}{Boolean indicating whether the palette should be reversed.}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE.}
}
\description{
Fill scale constructor for PoMS colours
}
